requirements.txt
# Currency ETL Project Dependencies

# HTTP requests
requests==2.32.5

# Data processing
pandas==2.3.3
numpy==2.3.4
pyarrow==22.0.0

# Database
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11

# Cloud storage
boto3==1.41.1

# Configuration
python-dotenv==1.2.1

# Utilities
python-dateutil==2.9.0
pytz==2025.2
tzdata==2025.2